requirements.txt
pydantic>=2.0.0
pydantic-ai[anthropic,google]>=0.1.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
streamlit>=1.30.0
scikit-learn>=1.4.0
nest-asyncio>=1.6.0
sniffio>=1.3.0
requests>=2.31.0
networkx>=3.2
supabase>=2.0.0
google-genai>=1.0.0
python-pptx>=0.6.0
filelock>=3.12.0
python-docx>=0.8.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
